Use "variant" to fix deprecation issue
It seems that the current version of the diffusers library does not
correctly handle the parameter `revision=fp16`. It suggests we use the
newer parameter instead: `variant=fp16`.

The code in the diffusers library seems to be written with the intention
of supporting either parameter (and warning users who use the deprecated
one), but it does not find the fp16 model weights correctly.

For details, see:
https://github.com/huggingface/diffusers/issues/5863#issuecomment-1827087787